x86-64: fix sysret stack alignment, strchr link gap, syscall sign-extension
Addresses a further round of Copilot review feedback on PR #350:

- bios/arch/x86_64/trapasm.S: the non-canonical-%rcx path called
  x86_64_bad_sysret() with %rsp%16==8, one push short of the alignment
  SysV requires immediately before a `call` -- the same class of bug
  trap1()/trap1_pexec() (miscasm.S) already needed a subq $8/addq $8
  pair for. My own comment justifying the omission was wrong: %rsp%16==8
  here is the residue of a just-*entered* callee (matching
  percpu.kernel_rsp's own "as if a return address had just been pushed"
  setup), not a caller ready to make a call of its own. Added the
  missing subq $8 (no matching addq needed: x86_64_bad_sysret() never
  returns).

- util/string.c: strlencpy()/strchr()/scasb()/expand_string() were
  gated to `#ifdef __arm__`, leaving x86-64 without a strchr()
  implementation at all. Latent today (pc-x86_64_defconfig enables
  neither CONF_WITH_ELF_LOADER nor CONF_WITH_PLUGGABLE_FS, both of
  which call it), but turning on either -- entirely plausible future
  configuration -- would produce an undefined reference at the PE link.
  Widened the guard to `#if defined(__arm__) || defined(__x86_64__)`,
  matching the pattern already used throughout this port; verified by
  temporarily enabling CONF_WITH_ELF_LOADER and confirming bdos/elfld.o
  links cleanly, then reverting that test config.

- include/bdosbind.h, cli/cmd.h: trap1()'s variadic `...` gives the
  compiler no real parameter type to convert an argument to, so a
  caller passing a plain 32-bit LONG/int scalar (e.g. Fseek()'s
  offset) only gets C's default variadic argument promotion (int stays
  int); the x86-64 backend's 32-bit register write for that argument
  then zero-extends it into the 64-bit register by ordinary ISA rules,
  turning a negative offset into a large positive one by the time
  osif() reads it back. Added explicit (long) casts to every
  non-opcode argument across both files' trap1()-based macros, which
  makes the C compiler perform the conversion at each call site
  correctly regardless of the caller's own literal/variable type --
  sign-extending a scalar, and losslessly reinterpreting a pointer,
  since long is this arch's native pointer width. A no-op on m68k/ARM
  (long and LONG are already the same width there), applied uniformly
  rather than only where it currently matters.

Rebuilt and boot-tested under QEMU/OVMF (pc-x86_64): reaches the same
gouser() milestone as before. Compile-tested virt-arm and virt-arm-cli
(no ARM QEMU available in this environment to also boot-test the
cli/cmd.h runtime path, unlike the x86-64 changes).

+++ cli/cmd.h
@@ -102,15 +102,23 @@ static __inline__ long cli_supexec_(long a)
 }
 #define Supexec(a) cli_supexec_((long)(a))
 
+/*
+ * (long), not (WORD)/(LONG), for every non-pointer argument here: see
+ * bdosbind.h's own comment on why trap1()'s variadic `...` needs an
+ * explicit widening cast at each call site to sign-extend correctly
+ * (rather than the zero-extension a narrower cast plus C's default
+ * variadic argument promotion produces on x86-64). (void *) casts are
+ * already full pointer width and need no change.
+ */
 #define jmp_gemdos_v(a)         trap1((int)(a))
-#define jmp_gemdos_w(a,b)       trap1((int)(a),(WORD)(b))
-#define jmp_gemdos_l(a,b)       trap1((int)(a),(LONG)(b))
+#define jmp_gemdos_w(a,b)       trap1((int)(a),(long)(b))
+#define jmp_gemdos_l(a,b)       trap1((int)(a),(long)(b))
 #define jmp_gemdos_p(a,b)       trap1((int)(a),(void*)(b))
-#define jmp_gemdos_ww(a,b,c)    trap1((int)(a),(WORD)(b),(WORD)(c))
-#define jmp_gemdos_pw(a,b,c)    trap1((int)(a),(void *)(b),(WORD)(c))
-#define jmp_gemdos_wlp(a,b,c,d) trap1((int)(a),(WORD)(b),(LONG)(c),(void *)(d))
-#define jmp_gemdos_wpp(a,b,c,d) trap1((int)(a),(WORD)(b),(void *)(c),(void *)(d))
-#define jmp_gemdos_pww(a,b,c,d) trap1((int)(a),(void *)(b),(WORD)(c),(WORD)(d))
+#define jmp_gemdos_ww(a,b,c)    trap1((int)(a),(long)(b),(long)(c))
+#define jmp_gemdos_pw(a,b,c)    trap1((int)(a),(void *)(b),(long)(c))
+#define jmp_gemdos_wlp(a,b,c,d) trap1((int)(a),(long)(b),(long)(c),(void *)(d))
+#define jmp_gemdos_wpp(a,b,c,d) trap1((int)(a),(long)(b),(void *)(c),(void *)(d))
+#define jmp_gemdos_pww(a,b,c,d) trap1((int)(a),(void *)(b),(long)(c),(long)(d))
 /* Pexec needs the 5-argument form; trap1_pexec handles the extra argument */
 #define jmp_gemdos_wppp(a,b,c,d,e) \
     trap1_pexec((short)(b),(const char *)(c),(const char *)(d),(const char *)(e))
